Add domain for user profile (#62)
* Add domain for user profile

* Add Supplier to AndroidApplicationService.
Remove demo code from MainNodePresenter and inti code to AndroidApplicationService.
Add generateKeyPairInProgress and createAndPublishInProgress properties.
Add createSimulatedDelay method to ClientUserProfileServiceFacade

Supplier is then used in DI and provides the services from the AndroidApplicationService when used by a lazy getter.
This hack is needed as AndroidApplicationService requires some constructor argumewnts which are not present as application startup but only after the mainActiviy is created.
We startup KOIN in the MainApplication.onCreate method.
We use the MainNodePresenter which gets called its onViewAttached once the mainActiviy is ready and then we can set the applicationService inside the supplier.

I don't like that unsafe lateinit setter strategy and have proposed an alterantive solution in my POC which would work with constructor injection. Maybe there are some hacks in KOIN to get the done in a different manner, but the main problem is that KOIN gets started at the MainApplication.onCreate and not on the MainActivity.onCreate. Only at that moment we have all the basic dependencies from the Android UI framework.
But even if we would move that to the MainActivity.onCreate, we would need to pass the properties to the module in some way.

* Rename AndroidModule to AndroidClientModule
Add DI config to AndroidClientModule
Remove ICreateProfilePresenter as there is only one instance and not expected a use case for multiple instances.
Add UserProfileRepository holding model and serviceFacade
Adjust Screen and Presenter.

Discussion:
To me the usage of the repository does not make much sense here.
The 2 relevant methods do not map into the CRUD pattern which seems to be the design philosophy for repositories.
I would suggest to use it only when teh CRUD pattern and persistence is used and appropriate. And to inject the model and the service facade to the presenter in the other cases.
Otherwise we just create a useless wrapper.
@rodvar: If you would implement the usage here in a different way, can you provide with the given use case an implementation example?

* Add "android.permission.INTERNET" entry to manifest for Android client

* Fix parameter in ApiRequestService. We pass the host instead of baseUrl.

On iOS simulator we need to use localhost, not 10.0.2.2

* Add IosClientModule
Use Dispatchers.Main instead of Dispatchers.IO as it seems iOS does not support that.

* Add equals and hashCode

* Remove redundant config for UserProfileRepository

* Remove unused model

* Add pruneAllBackups and readAllPersisted calls in AndroidApplicationService

* Add applySelectedUserProfile, getSelectedUserProfile and getUserIdentityIds methods.

In the CreateProfilePresenter at onLaunched there are commented out some different dev test modes;
- Use hasUserProfile to check if create user profile screen should be shown
- onGenerateKeyPair if no profile is present
- applySelectedUserProfile if one is present and we apply the data to the screen. This is only for dev testing. We have to display the selected user profile in the app in a TBD way.

Tested with all 3 apps and working.
Requires latest Bisq 2 main version and a running rest api.

* Move gradle dependencies to toml file

* Fix merge issue

* Fix merge issue
Go to next screen on onCreateAndPublishNewUserProfile completed.

* Move bisq2 compatible models to client.replicated_model

* Add clientModule, use DI for HttpClient

* Use class name for Logger.withTag
Remove unneeded coroutineScope.launch blocks

* Add ignoreUnknownKeys to json deserializer
Make Json  configs fields

* Add serviceFacade and model for bootstrap

* Refactor: Rename MainNodePresenter to NodeMainPresenter

* Add  serviceFacade and model for application bootstrap.

This only provides the domain part. It is not yet connected to the UI as its still unclear to me what patter to use for that.

When using the Client version one need to instantiate th ClientApplicationBootstrapFacade and call initialize(). Could be done in the presenter...

* Remove domain models and use return values instead

* Nits contribution

 - fix double call on view attached for the special case of the
   MainPresenter
 - remove double static configuration of bisq core jars (left in
   MainApplication override only)
 - provide single instance of android memory report service and adjust
   usages
 - cleanup/comment dead code
 - cleanup unused imports

* Apply code review suggestions

* Adjust to API changes

---------

Co-authored-by: Rodrigo Varela <rodvar@gmail.com>