Add create and take offer domains (#138)
* Refactoring: Rename network.bisq.mobile.android.node.domain to network.bisq.mobile.android.node.service .

Remove unused NodeGreetingRepository and Greeting classes

* Refactoring: Move serve packages from network.bisq.mobile.client to network.bisq.mobile.client.service.

* Refactoring: Move replicated_model package to domain. Rename to replicated

* Refactoring: Rename packages: offers, createOffer, takeOffer. Move take_offer to trade package

* Use new replicated data models.

We use 'VO' as postfix to the class name in the Bisq 2 model. For enums we use 'Enum' as postfix (better ideas welcome).

We need to distinguish the classes as both are used in the node module and to only distinguish by package name would be verbose and error prone.

VO seems to be a good compromise as they are not only data transfer objects but used as domain objects.

On Bisq 2 api side we use 'dto' as postfix as there they serve only the data transfer role. We do not use the original model classes anymore as it would cause too much pollution with Json serialisation annotations, swagger documentation and would add more risks with changes.

* Update websocket package with changes

* Add formatters, parsers, utils. Apply changes to existing services

* Refactoring: Move classes from utils package to domain/utils

* Add create offer and take offer domain.
Adjust existing ones to changes

* Add support for polymorphic data structures in json.
Adjust data models to changes.
Add decimalFormatter to PlatformDomainAbstractions

* Handle failures from loading persisted data more gracefully. At least during dev thats useful.

* Adjust to changes

* Add mappings between Bisq2 models to replicated data models (VOs)

* Add create offer and take offer service facades

* Adjust changes to services.

* Use new Mappings for getSelectedUserProfile

* Adopt to changes

* Adjust strings

* Refactor: Rename routes to match better Bisq2 domain language

* Add create and take offer screens and presenters

* UI Adjustments to changes

* Fix offer price display.
Change layout to match more closely Bisq 2

* - Onboarding navigation logic (#109)

---------

Co-authored-by: nostrbuddha <nostrbuddha@proton.me>